Add files via upload
This is the Python script for the question "Your company practices "acceptance sampling" on stock you receive from vendors. For a lot size of 150 units, you destructively test 20 randomly selected units. If more than 2 units do not conform to specifications, the entire lot of 150 units is rejected. What is the probability of rejecting the lot if the number of non-conforming units in the lot is given by the variable "bad"?"